feat(game): performance improvements & ajusted gap

+++ src/Bird.py
@@ -13,8 +13,8 @@ class Bird(pygame.sprite.Sprite):
         self.bird_image, self.bird = load_image_rect("assets/brainDefault.png", -1)
         self.bird_flap_image, self.bird_flap_rect = load_image_rect("assets/brainFlap.png", -1)
         self.bird_mid_flap_image, self.bird_mid_flap_rect = load_image_rect("assets/brainMidFlap.png", -1)
-        self.bird.center = (screen.get_width() / 2,
-                                  screen.get_height() / 2)
+        self.bird.center = (int(screen.get_width() / 2),
+                                  int(screen.get_height() / 2))
         self.first_jump = True
         self.image = self.bird_image
         self.rect = self.bird
